Major push to streamline new session creation.
o All functions are now exported using the safe_func_export --this
  functionality, whereby the script from which this functionality
  is called is parsed for function defs and each of those is
  exported. The call is placed after all functions are created.
  Calls that load individual functions are mostly removed, except
  in a few cases where we are bootstrapping early on in the process.

o Most aliases not set in the profiles were turned into functions. This
  make exporting to subshells possible for these definitions. We added
  a mechanism to look for a .alias file for each func file, but we
  are only using it for 'histfuncs' and 'perlfuncs'.

o Aliases set in profiles were left in the profiles since the profiles
  will get run on each invocation anyway.

o Loading of *funcs files is now skipped if the corresponding *reload
  function exists. Finding that function means that all the other
  related functions are loaded and we can skip the related script.

o A number of 'set' variables were 'export'ed so that subshells
  can access them.

o '_' in function names were changed to '-' in most cases.