The Great Autoloading Update - ip 1 70 18 156.dob1.bcpc.bloomberg.com (#223)
Converted all the functions in the *funcs files to autoloaded files in the functions directory.
Split out non exportables into .alias files.
Left the non-function exportables in .env files.
Starting up a new xterm is still a login operation, but now it's much faster.
Squashed 34 commits into one because many of the intermediate steps were "borken" anyway.
Did not convert the git functions to git scripts on PATH - that's for another day.