Cleaning up the git namespace and Darwinisms.
bin/git-is-clean
o Unused.

bin/git-retimestamp
o git-is => git-what-is

bin/git-sync-with-remote
o git-is => git-what-is

bin/pssearch
o Fixup for Darwin. Going with the short format all the time.

bin/xo
o Adjustments for Darwin.
o git-is => git-what-is

functions/git-cat-cl
o Deleting. Unused.

functions/git-check-hook
git-check-hook():
  o git-check-hook => git-hook-check

functions/git-dry-run-hook
git-dry-run-hook():
  o git-dry-run-hook => git-hook-dry-run

functions/git-go
git-go():
  o git-check-hook => git-hook-check

functions/git-hold
o git-is => git-what-is

functions/git-hook-check
git-hook-check():
  o Took the the guts from git-check-hook.

functions/git-hook-dry-run
git-hook-dry-run():
  o git-dry-run-hook => git-hook-dry-run

functions/git-hook-pull
git-hook-pull():
  o git-pull-hook => git-hook-pull

functions/git-hook-push
git-hook-push():
  o git-push-hook => git-hook-push

functions/git-is-committable
o Removed in favor of what-is

functions/git-is-deleted
o Removed in favor of what-is

functions/git-is-edited
o Removed in favor of what-is

functions/git-is-ignored
o Removed in favor of what-is

functions/git-is-mod
o Removed in favor of what-is

functions/git-is-modified
o Removed in favor of what-is

functions/git-is-staged
o Removed in favor of what-is

functions/git-is-unmodified
o Removed in favor of what-is

functions/git-is-unstaged
o Removed in favor of what-is

functions/git-is-untracked
o Removed in favor of what-is

functions/git-pull-hook
git-pull-hook():
  o git-pull-hook => git-hook-pull

functions/git-push-hook
git-push-hook():
  o git-push-hook => git-hook-push

functions/git-update-this-hook
git-update-this-hook():
  o Shellcheck.