Periodic
bin/cmd-echo
o Added -j for join so distinct arguments end up on one line.
o Skip all the coloring if the output is not a terminal; i.e., a log
  file.
o Added -i to force interactive mode and the colorization escape
  characters.

bin/cmd-yesno
o Throw the -i flag when getting the prompt string from cmd-echo.

bin/git-worktree
o script-yesno to cmd-yesno
o Added --no-reset flag so that multiple worktree calls in a loop won't
  waste time resetting the repo list.

functions/git-go-get-filename
o Removed in favor of a one-liner in the one place it's actually called.

functions/git-go-set
o git-go-get-filename call removed in favor of a one-liner in the one
  place it's actually called.